+++ backend/routes/auth.js
@@ -1,7 +1,29 @@
-const express = require("express");
+import express from "express";
+import rateLimit from "express-rate-limit";
+
 const router = express.Router();
 
-const auth = require("../services/auth");
+import * as auth from "../services/auth.js";
+
+const loginLimiter = rateLimit({
+  windowMs: (Number(process.env.ZU_LOGIN_LIMIT_WINDOW) || 30) * 60 * 1000, // 30 minutes
+  max: Number(process.env.ZU_LOGIN_LIMIT_ATTEMPTS) || 50, // limit each IP to 50 requests per windowMs
+  message: {
+    status: 429,
+    error: "tooManyAttempts",
+  },
+});
+
+const loginLimiterWrapper = (req, res, next) => {
+  if (
+    process.env.NODE_ENV === "production" &&
+    process.env.ZU_LOGIN_LIMIT === "true"
+  ) {
+    return loginLimiter(req, res, next);
+  } else {
+    return next();
+  }
+};
 
 router.get("/login", async function (req, res) {
   if (process.env.ZU_DISABLE_AUTH === "true") {
@@ -11,7 +33,7 @@ router.get("/login", async function (req, res) {
   }
 });
 
-router.post("/login", async function (req, res) {
+router.post("/login", loginLimiterWrapper, async function (req, res) {
   if (req.body.username && req.body.password) {
     auth.authorize(req.body.username, req.body.password, function (err, user) {
       if (user) {
@@ -27,4 +49,4 @@ router.post("/login", async function (req, res) {
   }
 });
 
-module.exports = router;
+export default router;