+++ backend/services/auth.js
@@ -1,25 +1,23 @@
-const db = require("../utils/db");
-const verifyHash = require("pbkdf2-wrapper/verifyHash");
+import { db } from "../utils/db.js";
+import verifyHash from "pbkdf2-wrapper/verifyHash.js";
 
-exports.authorize = authorize;
-async function authorize(username, password, callback) {
+export async function authorize(username, password, callback) {
   try {
     var users = await db.get("users");
   } catch (err) {
     throw err;
   }
   const user = users.find({ username: username });
-  if (!user.value()) return callback(new Error("Cannot find user"));
+  if (!user.value()) return callback(new Error("logInFailed")); // If return "user not found" someone can do a user listing
   const verified = await verifyHash(password, user.value()["password_hash"]);
   if (verified) {
     return callback(null, user.value());
   } else {
-    return callback(new Error("Invalid password"));
+    return callback(new Error("logInFailed"));
   }
 }
 
-exports.isAuthorized = isAuthorized;
-async function isAuthorized(req, res, next) {
+export async function isAuthorized(req, res, next) {
   if (process.env.ZU_DISABLE_AUTH === "true") {
     next();
   } else {