+++ frontend/src/components/HomeLoggedOut/HomeLoggedOut.jsx
@@ -1,19 +1,38 @@
+import { useEffect } from "react";
 import { Grid, Typography } from "@material-ui/core";
 import { useLocalStorage } from "react-use";
-import axios from "axios";
 import { useHistory } from "react-router-dom";
+
+import { useTranslation } from "react-i18next";
+
+import axios from "axios";
 
 function HomeLoggedOut() {
   const [, setLoggedIn] = useLocalStorage("loggedIn", false);
   const [, setToken] = useLocalStorage("token", null);
+  const [, setDisableAuth] = useLocalStorage("disableAuth", false);
   const history = useHistory();
-  axios.get("/auth/login").then(function (response) {
-    if (!response.data.enabled) {
-      setLoggedIn(true);
-      setToken("");
-      history.go(0);
+
+  useEffect(() => {
+    async function fetchData() {
+      axios
+        .get("/auth/login", { withCredentials: true })
+        .then(function (response) {
+          if (!response.data.enabled) {
+            setLoggedIn(true);
+            setDisableAuth(true);
+            setToken("");
+            history.go(0);
+          } else {
+            setDisableAuth(false);
+          }
+        });
     }
-  });
+    fetchData();
+  }, [history, setDisableAuth, setLoggedIn, setToken]);
+
+  const { t, i18n } = useTranslation();
+
   return (
     <Grid
       container
@@ -27,14 +46,11 @@ function HomeLoggedOut() {
     >
       <Grid item xs={10}>
         <Typography variant="h5">
-          <span>
-            ZeroUI - ZeroTier Controller Web UI - is a web user interface for a
-            self-hosted ZeroTier network controller.
-          </span>
+          <span>{t("zerouiDesc")}</span>
         </Typography>
 
         <Typography>
-          <span>Please Log In to continue</span>
+          <span>{t("loginToContinue")}</span>
         </Typography>
       </Grid>
     </Grid>