+++ frontend/src/components/NetworkMembers/NetworkMembers.jsx
@@ -14,11 +14,14 @@ import DataTable from "react-data-table-component";
 import { useParams } from "react-router-dom";
 import API from "utils/API";
 import { parseValue, replaceValue, setValue } from "utils/ChangeHelper";
+import { formatDistance } from "date-fns";
 import AddMember from "./components/AddMember";
 import DeleteMember from "./components/DeleteMember";
 import ManagedIP from "./components/ManagedIP";
 import MemberName from "./components/MemberName";
 import MemberSettings from "./components/MemberSettings";
+
+import { useTranslation } from "react-i18next";
 
 function NetworkMembers({ network }) {
   const { nwid } = useParams();
@@ -45,6 +48,8 @@ function NetworkMembers({ network }) {
     console.log("Action:", req);
   };
 
+  const { t, i18n } = useTranslation();
+
   const handleChange =
     (member, key1, key2 = null, mode = "text", id = null) =>
     (event) => {
@@ -66,7 +71,7 @@ function NetworkMembers({ network }) {
   const columns = [
     {
       id: "auth",
-      name: "Authorized",
+      name: t("authorized"),
       minWidth: "80px",
       cell: (row) => (
         <Checkbox
@@ -78,7 +83,7 @@ function NetworkMembers({ network }) {
     },
     {
       id: "address",
-      name: "Address",
+      name: t("address"),
       minWidth: "150px",
       cell: (row) => (
         <Typography variant="body2">{row.config.address}</Typography>
@@ -86,52 +91,52 @@ function NetworkMembers({ network }) {
     },
     {
       id: "name",
-      name: "Name / Description",
+      name: t("name") + "/" + t("description"),
       minWidth: "250px",
       cell: (row) => <MemberName member={row} handleChange={handleChange} />,
     },
     {
       id: "ips",
-      name: "Managed IPs",
+      name: t("managedIPs"),
       minWidth: "220px",
       cell: (row) => <ManagedIP member={row} handleChange={handleChange} />,
     },
     {
-      id: "status",
-      name: "Peer status",
+      id: "lastSeen",
+      name: t("lastSeen"),
       minWidth: "100px",
       cell: (row) =>
-        row.online === 0 ? (
-          <Typography color="error">OFFLINE</Typography>
-        ) : row.online === 1 ? (
-          <Typography style={{ color: "#008000" }}>
-            {"ONLINE (v" +
-              row.config.vMajor +
-              "." +
-              row.config.vMinor +
-              "." +
-              row.config.vRev +
-              ")"}
+        row.online === 1 ? (
+          <Typography style={{ color: "#008000" }}>{"ONLINE"}</Typography>
+        ) : row.controllerId === row.config.address ? (
+          <Typography style={{ color: "#c5e31e" }}>{"CONTROLLER"}</Typography>
+        ) : row.online === 0 ? (
+          <Typography color="error">
+            {row.lastOnline !== 0
+              ? formatDistance(row.lastOnline, row.clock, {
+                  includeSeconds: false,
+                  addSuffix: true,
+                })
+              : "OFFLINE"}
           </Typography>
         ) : (
-          <Typography style={{ color: "#f1c232" }}>
-            {"RELAYED (v" +
-              row.config.vMajor +
-              "." +
-              row.config.vMinor +
-              "." +
-              row.config.vRev +
-              ")"}
-          </Typography>
+          <Typography style={{ color: "#f1c232" }}>{"RELAYED"}</Typography>
         ),
     },
     {
       id: "physicalip",
-      name: "Physical IP / Latency",
+      name: t("version") + " / " + t("physIp") + " / " + t("latency"),
       minWidth: "220px",
       cell: (row) =>
         row.online === 1 ? (
           <p>
+            {"v" +
+              row.config.vMajor +
+              "." +
+              row.config.vMinor +
+              "." +
+              row.config.vRev}
+            <br />
             {row.physicalAddress + "/" + row.physicalPort}
             <br />
             {"(" + row.latency + " ms)"}
@@ -142,7 +147,7 @@ function NetworkMembers({ network }) {
     },
     {
       id: "delete",
-      name: "",
+      name: t("settings"),
       minWidth: "50px",
       right: true,
       cell: (row) => (
@@ -161,7 +166,7 @@ function NetworkMembers({ network }) {
   return (
     <Accordion defaultExpanded={true}>
       <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-        <Typography>Members</Typography>
+        <Typography>{t("member", { count: members.length })}</Typography>
       </AccordionSummary>
       <AccordionDetails>
         <Grid container direction="column" spacing={3}>
@@ -181,14 +186,13 @@ function NetworkMembers({ network }) {
                 spacing={0}
                 direction="column"
                 alignItems="center"
-                justify="center"
+                justifyContent="center"
                 style={{
                   minHeight: "50vh",
                 }}
               >
                 <Typography variant="h6" style={{ padding: "10%" }}>
-                  No devices have joined this network. Use the app on your
-                  devices to join <b>{nwid}</b>.
+                  {t("noDevices")} <b>{nwid}</b>.
                 </Typography>
               </Grid>
             )}