+++ frontend/src/routes/Network/Network.jsx
@@ -11,7 +11,10 @@ import { useLocalStorage } from "react-use";
 import API from "utils/API";
 import useStyles from "./Network.styles";
 
+import { useTranslation } from "react-i18next";
+
 function Network() {
+  const { t, i18n } = useTranslation();
   const { nwid } = useParams();
   const [loggedIn] = useLocalStorage("loggedIn", false);
   const [network, setNetwork] = useState({});
@@ -39,10 +42,12 @@ function Network() {
   if (loggedIn) {
     return (
       <>
-        <Link color="inherit" component={RouterLink} to="/" underline="none">
-          <ArrowBackIcon className={classes.backIcon}></ArrowBackIcon>
-          Networks
-        </Link>
+        <div className={classes.breadcrumbs}>
+          <Link color="inherit" component={RouterLink} to="/" underline="none">
+            <ArrowBackIcon className={classes.backIcon}></ArrowBackIcon>
+            {t("network", { count: 2 })}
+          </Link>
+        </div>
         <div className={classes.container}>
           {network["config"] && (
             <>
@@ -71,9 +76,7 @@ function Network() {
         }}
       >
         <Grid item xs={10}>
-          <Typography variant="h5">
-            You are not authorized. Please Log In
-          </Typography>
+          <Typography variant="h5">{t("notAuthorized")}</Typography>
         </Grid>
       </Grid>
     );